second draft of sort by distance with loading component
