+++ src/components/location-finder/Locations/Locations.js
@@ -1,15 +1,16 @@
-import React from 'react';
+import React, { Fragment } from 'react';
+// Apollo
 import gql from 'graphql-tag';
 import { useQuery } from '@apollo/react-hooks';
-import Location from './../Location';
 import { LocationsQuery as LOCATIONS_QUERY } from './Locations.gql';
-import * as DS from '@nypl/design-system-react-components';
 // Map
 import Map from './../Map';
 // Redux
 import { useDispatch, useSelector } from 'react-redux';
 import { setSearchQuery, setMapPosition, setLocationInfoWindowId, setAutoSuggestInputValue } from './../../../redux/actions';
-
+// Components
+import * as DS from '@nypl/design-system-react-components';
+import Location from './../Location';
 import Skeleton from 'react-loading-skeleton';
 
 function Locations() {
@@ -22,6 +23,7 @@ function Locations() {
 
   const dispatch = useDispatch();
 
+  // Apollo
   const searchGeoLat = searchQueryGeoLat ? searchQueryGeoLat : 40.7532;
   const searchGeoLng = searchQueryGeoLng ? searchQueryGeoLng : -73.9822;
 
@@ -39,9 +41,9 @@ function Locations() {
 
     return (
       <Skeleton
-        height={75}
+        height={20}
         count={20}
-        duration={2}
+        duration={1.2}
       />
     );
   }
@@ -52,64 +54,12 @@ function Locations() {
     );
   }
 
-  function onClearSearchTerms(e) {
-    e.preventDefault();
-    console.log('Clear all search terms!');
-
-    dispatch(setSearchQuery({
-      searchQuery: '',
-      searchQueryGeoLat: '',
-      searchQueryGeoLng: ''
-    }));
-
-    const defaultCenter = {
-      lat: 40.7532,
-      lng: -73.9822
-    };
-
-    dispatch(setMapPosition({
-      mapCenter: defaultCenter,
-      mapZoom: 12
-    }));
-
-    // Dispatch to reset the location id for info window.
-    dispatch(setLocationInfoWindowId(null));
-
-    // Clear auto suggest input.
-    dispatch(setAutoSuggestInputValue(''));
-  }
-
   return (
-    <div className='locations'>
-      <div className='row'>
-        <div className='column locations__list'>
-
-          {searchQuery ? (
-            <div>
-              Showing all locations near <strong>{searchQuery}</strong>
-              <br />
-              <DS.Link
-                href="#"
-                onClick={onClearSearchTerms}
-              >
-                Clear all search terms
-              </DS.Link>
-            </div>
-          ) : (
-            null
-          )}
-
-          <div>
-            {data.allLocations.map((location) => (
-              <Location key={location.id} location={location} />
-            ))}
-          </div>
-        </div>
-        <div className='column locations__map'>
-          <Map locations={data.allLocations} />
-        </div>
-      </div>
-    </div>
+    <Fragment>
+      {data.allLocations.map((location) => (
+        <Location key={location.id} location={location} />
+      ))}
+    </Fragment>
   );
 }
