+++ src/components/location-finder/Map/Locations.gql
@@ -1,0 +1,18 @@
+query LocationsQuery {
+  allLocations {
+    id
+    name
+    status
+    address_line1
+    address_line2
+    locality
+    administrative_area
+    postal_code
+    phone
+    wheelchairAccess
+    geoLocation {
+      lat
+      lng
+    }
+  }
+}
