+++ src/pages/location-finder/index.js
@@ -7,6 +7,8 @@ import { compose } from 'redux';
 import Layout from './../../components/shared/layouts/Main';
 import SearchHeader from './../../components/location-finder/SearchHeader';
 import Locations from './../../components/location-finder/Locations/Locations';
+import SearchResultsDetails from './../../components/location-finder/SearchResultsDetails';
+import Map from './../../components/location-finder/Map';
 import Head from 'next/head';
 
 function LocationFinder() {
@@ -20,7 +22,17 @@ function LocationFinder() {
         <SearchHeader />
       </div>
       <div className="content-primary">
-        <Locations />
+        <div className='locations'>
+          <SearchResultsDetails />
+          <div className='row'>
+            <div className='column locations__list'>
+              <Locations />
+            </div>
+            <div className='column locations__map'>
+              <Map />
+            </div>
+          </div>
+        </div>
       </div>
       <div className="content-bottom">
         Content Bottom
